--- a/packages/node/test/node_test.js
+++ b/packages/node/test/node_test.js
@@ -39,6 +39,11 @@ test('valid preset production', t => {
   t.deepEqual(config.resolve.extensions, expectedExtensions);
   t.is(config.optimization, undefined);
   t.is(config.devServer, undefined);
+  t.deepEqual(config.stats, {
+    children: false,
+    entrypoints: false,
+    modules: false
+  });
 
   // NODE_ENV/command specific
   t.is(config.devtool, 'source-map');
@@ -58,6 +63,11 @@ test('valid preset development', t => {
   t.deepEqual(config.resolve.extensions, expectedExtensions);
   t.is(config.optimization, undefined);
   t.is(config.devServer, undefined);
+  t.deepEqual(config.stats, {
+    children: false,
+    entrypoints: false,
+    modules: false
+  });
 
   // NODE_ENV/command specific
   t.is(config.devtool, 'inline-sourcemap');