--- a/packages/web/index.js
+++ b/packages/web/index.js
@@ -191,6 +191,12 @@ module.exports = (neutrino, opts = {}) => {
           .end()
         .end()
       .end()
+    // The default output is too noisy, particularly with multiple entrypoints.
+    .stats({
+      children: false,
+      entrypoints: false,
+      modules: false
+    })
     .when(neutrino.config.module.rules.has('lint'), () => {
       neutrino.use(loaderMerge('lint', 'eslint'), {
         envs: ['browser', 'commonjs']