This adjusts the console output to be less verbose, for parity with how it was prior to the switch to native CLIs in #852.

In some cases the output has been made to be even more concise than in Neutrino 8, since it's now possible for users to adjust the output via CLI flags, therefore less important to satisfy all use-cases out of the box. For that reason the `debug` handling has also been removed.

The output when using webpack-dev-server now includes the duration, which makes it much easier to compare incremental build times.

The `performance.hints(false)` of `@neutrinojs/node` has been removed since it's redundant as of webpack 4.2.0:
https://github.com/webpack/webpack/commit/c65fb74a2606631a1d681ab1af0fd53c0d922893

Fixes #897.