@edmorley 2018-09-04
Example log output using `@neutrinojs/web` and `@neutrinojs/karma` project created using create-project...

### yarn build

Before:

```
$ yarn build
yarn run v1.9.4
$ webpack --mode production
Hash: 7f7e5549edcaaf3cee33
Version: webpack 4.17.1
Time: 1107ms
Built at: 09/04/2018 12:59:14 AM
              Asset       Size  Chunks             Chunk Names
  index.64bb17af.js  227 bytes       0  [emitted]  index
runtime.b2f53520.js   1.42 KiB       1  [emitted]  runtime
         index.html  345 bytes          [emitted]
Entrypoint index = runtime.b2f53520.js index.64bb17af.js
[0] multi ./src/index 28 bytes {0} [built]
[1] ./src/index.js 220 bytes {0} [built]
Child html-webpack-plugin for "index.html":
     1 asset
    Entrypoint undefined = index.html
    [0] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!../neutrino/packages/html-template/template.ejs 573 bytes {0} [built]
    [2] (webpack)/buildin/global.js 489 bytes {0} [built]
    [3] (webpack)/buildin/module.js 497 bytes {0} [built]
        + 1 hidden module
Done in 4.13s.
```

After:

```
$ yarn build
yarn run v1.9.4
$ webpack --mode production
Hash: 7f7e5549edcaaf3cee33
Version: webpack 4.17.1
Time: 1077ms
Built at: 09/04/2018 12:55:08 AM
              Asset       Size  Chunks             Chunk Names
  index.64bb17af.js  227 bytes       0  [emitted]  index
runtime.b2f53520.js   1.42 KiB       1  [emitted]  runtime
         index.html  345 bytes          [emitted]
Done in 4.01s.
```

### yarn start

Before:

```
$ yarn start
yarn run v1.9.4
$ webpack-dev-server --mode development
i ｢wds｣: Project is running at http://localhost:5000/
i ｢wds｣: webpack output is served from /
i ｢wds｣: Content not from webpack is served from C:\Users\Ed\src\test-project\src
i ｢wds｣: 404s will fallback to /index.html
i ｢wdm｣: Built at: 09/04/2018 12:58:48 AM
Entrypoint index = runtime.js vendors~index.js index.js
i ｢wdm｣: Compiled successfully.
i ｢wdm｣: Compiling...
i ｢wdm｣: Built at: 09/04/2018 12:58:53 AM
Entrypoint index = runtime.js vendors~index.js index.js
i ｢wdm｣: Compiled successfully.
```

After:

```
$ yarn start
yarn run v1.9.4
$ webpack-dev-server --mode development
i ｢wds｣: Project is running at http://localhost:5000/
i ｢wds｣: webpack output is served from /
i ｢wds｣: Content not from webpack is served from C:\Users\Ed\src\test-project\src
i ｢wds｣: 404s will fallback to /index.html
i ｢wdm｣: Time: 1681ms
i ｢wdm｣: Compiled successfully.
i ｢wdm｣: Compiling...
i ｢wdm｣: Time: 99ms
i ｢wdm｣: Compiled successfully.
```

### yarn test

Before:

```
$ yarn test
yarn run v1.9.4
$ karma start --single-run

START:
i ｢wdm｣:
i ｢wdm｣: Compiled successfully.
i ｢wdm｣: Compiling...
i ｢wdm｣:    7 modules
i ｢wdm｣: Compiled successfully.
04 09 2018 00:58:32.615:INFO [karma]: Karma v3.0.0 server started at http://0.0.0.0:9876/
04 09 2018 00:58:32.620:INFO [launcher]: Launching browser ChromeHeadless with unlimited concurrency
04 09 2018 00:58:32.644:INFO [launcher]: Starting browser ChromeHeadless
04 09 2018 00:58:33.444:INFO [HeadlessChrome 0.0.0 (Windows 10 0.0.0)]: Connected on socket OsAafN9KbIsDOOX2AAAA with id 41761616
  simple
    √ should be sane

Finished in 0.012 secs / 0 secs @ 00:58:33 GMT+0100 (British Summer Time)

SUMMARY:
√ 1 test completed
Done in 6.32s.
```

After:

```
$ yarn test
yarn run v1.9.4
$ karma start --single-run

START:
i ｢wdm｣: Time: 96ms
i ｢wdm｣: Compiled successfully.
i ｢wdm｣: Compiling...
i ｢wdm｣: Time: 1111ms
i ｢wdm｣: Compiled successfully.
04 09 2018 00:57:57.602:INFO [karma]: Karma v3.0.0 server started at http://0.0.0.0:9876/
04 09 2018 00:57:57.607:INFO [launcher]: Launching browser ChromeHeadless with unlimited concurrency
04 09 2018 00:57:57.627:INFO [launcher]: Starting browser ChromeHeadless
04 09 2018 00:57:58.307:INFO [HeadlessChrome 0.0.0 (Windows 10 0.0.0)]: Connected on socket C7OIbp_biXRkSHf3AAAA with id 73455848
  simple
    √ should be sane

Finished in 0.009 secs / 0 secs @ 00:57:58 GMT+0100 (British Summer Time)

SUMMARY:
√ 1 test completed
Done in 6.23s.
```

---

Even more extreme is when building a project with multiple entrypoints and a large dependency graph (such as Treeherder):

### yarn build

Before:

```
$ yarn build
yarn run v1.9.4
$ node ./node_modules/webpack/bin/webpack.js --mode production
Hash: e4a0296891796b7f3737
Version: webpack 4.17.1
Time: 29951ms
Built at: 08/30/2018 9:30:15 PM
                                 Asset       Size      Chunks             Chunk Names
                     0.d6e4c36b.js.map   90 bytes           0  [emitted]
      fontawesome-webfont.674f50d2.eot    162 KiB              [emitted]
     fontawesome-webfont.fee66e71.woff   95.7 KiB              [emitted]
      fontawesome-webfont.b06871f2.ttf    162 KiB              [emitted]
      fontawesome-webfont.912ec66d.svg    434 KiB              [emitted]
              dancing_cat.fa5552a5.gif     58 KiB              [emitted]
          treeherder-logo.3df97cff.png   8.45 KiB              [emitted]
                        0.dbd1207b.css   36.5 KiB           0  [emitted]
                         0.d6e4c36b.js  121 bytes           0  [emitted]
                   runtime.b2f53520.js   1.46 KiB           1  [emitted]  runtime
                        2.1224ffba.css    138 KiB           2  [emitted]
                         2.79309b4a.js  121 bytes           2  [emitted]
                         3.fdfa32ff.js   93.2 KiB           3  [emitted]
                         4.9365147e.js   85.3 KiB           4  [emitted]
                         5.bc14dd84.js    104 KiB           5  [emitted]
                    index.350cffb4.css   53.8 KiB   6, 13, 14  [emitted]  index
                     index.282ec1ba.js   1.61 MiB   6, 13, 14  [emitted]  index
                logviewer.660855ce.css   10.8 KiB       7, 13  [emitted]  logviewer
                 logviewer.2d5ceb27.js    628 KiB       7, 13  [emitted]  logviewer
                userguide.7c2cb730.css   23.4 KiB           8  [emitted]  userguide
                 userguide.b5d92f8e.js   4.33 KiB           8  [emitted]  userguide
                    login.37e77cf5.css  119 bytes           9  [emitted]  login
                     login.073e0707.js   21.8 KiB           9  [emitted]  login
                 testview.c1968046.css   2.96 KiB          10  [emitted]  testview
                  testview.b69ff139.js    229 KiB          10  [emitted]  testview
                     perf.51ff12c0.css   33.8 KiB  11, 13, 14  [emitted]  perf
                      perf.c570f5a6.js    1.6 MiB  11, 13, 14  [emitted]  perf
    intermittent-failures.2b7982b3.css   30.1 KiB       12, 5  [emitted]  intermittent-failures
     intermittent-failures.fe1f54b2.js   1.06 MiB       12, 5  [emitted]  intermittent-failures
                        13.f6b18392.js    178 KiB          13  [emitted]
                        14.e03f7220.js    105 KiB          14  [emitted]
                    0.dbd1207b.css.map   52.3 KiB           0  [emitted]
    fontawesome-webfont.af7ae505.woff2   75.4 KiB              [emitted]
               runtime.b2f53520.js.map   7.76 KiB           1  [emitted]  runtime
                    2.1224ffba.css.map    173 KiB           2  [emitted]
                     2.79309b4a.js.map   90 bytes           2  [emitted]
                     3.fdfa32ff.js.map    186 KiB           3  [emitted]
                     4.9365147e.js.map    423 KiB           4  [emitted]
                     5.bc14dd84.js.map    258 KiB           5  [emitted]
                index.350cffb4.css.map   70.6 KiB   6, 13, 14  [emitted]  index
                 index.282ec1ba.js.map   6.96 MiB   6, 13, 14  [emitted]  index
            logviewer.660855ce.css.map   14.6 KiB       7, 13  [emitted]  logviewer
             logviewer.2d5ceb27.js.map   3.68 MiB       7, 13  [emitted]  logviewer
            userguide.7c2cb730.css.map   30.3 KiB           8  [emitted]  userguide
             userguide.b5d92f8e.js.map   6.03 KiB           8  [emitted]  userguide
                login.37e77cf5.css.map  239 bytes           9  [emitted]  login
                 login.073e0707.js.map   51.8 KiB           9  [emitted]  login
             testview.c1968046.css.map    4.1 KiB          10  [emitted]  testview
              testview.b69ff139.js.map    919 KiB          10  [emitted]  testview
                 perf.51ff12c0.css.map   45.4 KiB  11, 13, 14  [emitted]  perf
                  perf.c570f5a6.js.map   7.32 MiB  11, 13, 14  [emitted]  perf
intermittent-failures.2b7982b3.css.map   39.8 KiB       12, 5  [emitted]  intermittent-failures
 intermittent-failures.fe1f54b2.js.map   4.06 MiB       12, 5  [emitted]  intermittent-failures
                    13.f6b18392.js.map    1.6 MiB          13  [emitted]
                    14.e03f7220.js.map    267 KiB          14  [emitted]
                            index.html   1.97 KiB              [emitted]
                        logviewer.html   5.06 KiB              [emitted]
                        userguide.html    9.9 KiB              [emitted]
                            login.html  558 bytes              [emitted]
                         testview.html  614 bytes              [emitted]
                             perf.html   3.13 KiB              [emitted]
            intermittent-failures.html  646 bytes              [emitted]
                            robots.txt   26 bytes              [emitted]
                       contribute.json  970 bytes              [emitted]
                          revision.txt  252 bytes              [emitted]
Entrypoint index [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 4.9365147e.js 4.9365147e.js.map index.350cffb4.css index.282ec1ba.js index.350cffb4.css.map index.282ec1ba.js.map
Entrypoint logviewer [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 4.9365147e.js 4.9365147e.js.map logviewer.660855ce.css logviewer.2d5ceb27.js logviewer.660855ce.css.map logviewer.2d5ceb27.js.map
Entrypoint userguide [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 4.9365147e.js 4.9365147e.js.map 13.f6b18392.js 13.f6b18392.js.map userguide.7c2cb730.css userguide.b5d92f8e.js userguide.7c2cb730.css.map userguide.b5d92f8e.js.map
Entrypoint login [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 5.bc14dd84.js 5.bc14dd84.js.map 14.e03f7220.js 14.e03f7220.js.map login.37e77cf5.css login.073e0707.js login.37e77cf5.css.map login.073e0707.js.map
Entrypoint testview [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 5.bc14dd84.js 5.bc14dd84.js.map testview.c1968046.css testview.b69ff139.js testview.c1968046.css.map testview.b69ff139.js.map
Entrypoint perf [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 4.9365147e.js 4.9365147e.js.map perf.51ff12c0.css perf.c570f5a6.js perf.51ff12c0.css.map perf.c570f5a6.js.map
Entrypoint intermittent-failures [big] = runtime.b2f53520.js runtime.b2f53520.js.map 0.dbd1207b.css 0.d6e4c36b.js 0.dbd1207b.css.map 0.d6e4c36b.js.map 2.1224ffba.css 2.79309b4a.js 2.1224ffba.css.map 2.79309b4a.js.map 3.fdfa32ff.js 3.fdfa32ff.js.map 4.9365147e.js 4.9365147e.js.map intermittent-failures.2b7982b3.css intermittent-failures.fe1f54b2.js intermittent-failures.2b7982b3.css.map intermittent-failures.fe1f54b2.js.map
[574] multi ./ui/entry-index.js 28 bytes {6} [built]
[591] ./ui/job-view/JobView.jsx 11.2 KiB {6} [built]
[680] multi ./ui/entry-logviewer.js 28 bytes {7} [built]
[683] multi ./ui/entry-userguide.js 28 bytes {8} [built]
[686] multi ./ui/entry-login.jsx 28 bytes {9} [built]
[687] ./ui/entry-login.jsx 268 bytes {9} [built]
[690] multi ./ui/test-view/index.jsx 28 bytes {10} [built]
[691] ./ui/test-view/index.jsx 453 bytes {10} [built]
[705] multi ./ui/entry-perf.js 28 bytes {11} [built]
[727] multi ./ui/intermittent-failures/index.jsx 28 bytes {12} [built]
[745] ./ui/entry-perf.js + 13 modules 156 KiB {11} [built]
      | ./ui/js/controllers/perf/compare.js 38 KiB [built]
      | ./ui/entry-perf.js 1.3 KiB [built]
      | ./ui/js/models/perf/performance_framework.js 362 bytes [built]
      | ./ui/js/models/perf/alerts.js 13.4 KiB [built]
      | ./ui/js/services/perf/math.js 2 KiB [built]
      | ./ui/js/services/perf/compare.js 14.9 KiB [built]
      | ./ui/js/perf.js 462 bytes [built]
      | ./ui/js/models/perf/issue_tracker.js 644 bytes [built]
      | ./ui/js/controllers/perf/alerts.js 21.1 KiB [built]
      | ./ui/js/controllers/perf/dashboard.js 14.6 KiB [built]
      | ./ui/js/perfapp.js 4.63 KiB [built]
      | ./ui/js/controllers/perf/graphs.js 38.7 KiB [built]
      |     + 2 hidden modules
[746] ./ui/entry-index.js + 5 modules 50.6 KiB {6} [built]
      | ./ui/entry-index.js 1.07 KiB [built]
      | ./ui/js/treeherder_app.js 2.1 KiB [built]
      | ./ui/js/services/buildapi.js 2.12 KiB [built]
      | ./ui/js/models/resultsets_store.js 32.5 KiB [built]
      | ./ui/js/controllers/main.js 12.3 KiB [built]
      | ./ui/models/runnableJob.js 331 bytes [built]
[747] ./ui/intermittent-failures/index.jsx + 4 modules 4.08 KiB {12} [built]
      | ./ui/intermittent-failures/index.jsx 382 bytes [built]
      | ./ui/intermittent-failures/redux/store.js 214 bytes [built]
      | ./ui/intermittent-failures/redux/root_reducer.js 664 bytes [built]
      | ./ui/intermittent-failures/redux/reducers.js 2.33 KiB [built]
      |     + 1 hidden module
[748] ./ui/entry-logviewer.js + 4 modules 12.9 KiB {7} [built]
      | ./ui/entry-logviewer.js 623 bytes [built]
      | ./ui/js/logviewer.js 1.02 KiB [built]
      | ./ui/js/directives/treeherder/log_viewer_steps.js 1.08 KiB [built]
      | ./ui/js/controllers/logviewer.js 8.88 KiB [built]
      |     + 1 hidden module
[749] ./ui/entry-userguide.js + 2 modules 1.69 KiB {8} [built]
      | ./ui/entry-userguide.js 449 bytes [built]
      | ./ui/js/userguide.js 835 bytes [built]
      | ./ui/js/controllers/userguide.js 413 bytes [built]
    + 749 hidden modules
Child HtmlWebpackCompiler:
     5 assets
    Entrypoint HtmlWebpackPlugin_0 = __child-HtmlWebpackPlugin_0
    Entrypoint HtmlWebpackPlugin_1 = __child-HtmlWebpackPlugin_1
    Entrypoint HtmlWebpackPlugin_2 = __child-HtmlWebpackPlugin_2
    Entrypoint HtmlWebpackPlugin_3 = __child-HtmlWebpackPlugin_3
    Entrypoint HtmlWebpackPlugin_4 = __child-HtmlWebpackPlugin_4
     [0] ./ui/img/tree_open.png 753 bytes {0} {2} [built]
     [1] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!./ui/index.html 1.2 KiB {0} [built]
     [2] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!./ui/logviewer.html 5.34 KiB {1} [built]
     [3] ./ui/img/logviewerIcon.png 1.08 KiB {1} [built]
     [4] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!./ui/userguide.html 9.32 KiB {2} [built]
     [5] ./ui/img/logviewerIconHelp.svg 3.02 KiB {2} [built]
     [6] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!../neutrino/packages/html-template/template.ejs 573 bytes {3} [built]
     [8] (webpack)/buildin/global.js 489 bytes {3} [built]
     [9] (webpack)/buildin/module.js 497 bytes {3} [built]
    [10] ../neutrino/node_modules/html-webpack-plugin/lib/loader.js!./ui/perf.html 2.98 KiB {4} [built]
    [11] ./ui/img/line_chart.png 481 bytes {4} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!node_modules/bootstrap/dist/css/bootstrap.min.css:
    Entrypoint mini-css-extract-plugin = *
       2 modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!node_modules/font-awesome/css/font-awesome.css:
                                 Asset      Size  Chunks            Chunk Names
      fontawesome-webfont.674f50d2.eot   162 KiB          [emitted]
    fontawesome-webfont.af7ae505.woff2  75.4 KiB          [emitted]
     fontawesome-webfont.fee66e71.woff  95.7 KiB          [emitted]
      fontawesome-webfont.b06871f2.ttf   162 KiB          [emitted]
      fontawesome-webfont.912ec66d.svg   434 KiB          [emitted]
    Entrypoint mini-css-extract-plugin = *
       9 modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!node_modules/metrics-graphics/dist/metricsgraphics.css:
    Entrypoint mini-css-extract-plugin = *
       2 modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!node_modules/react-day-picker/lib/style.css:
    Entrypoint mini-css-extract-plugin = *
       2 modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!node_modules/react-table/react-table.css:
    Entrypoint mini-css-extract-plugin = *
       2 modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/intermittent-failures.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/intermittent-failures.css 2.08 KiB {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/login.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/login.css 251 bytes {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/logviewer.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/logviewer.css 3.3 KiB {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/perf.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/perf.css 9.96 KiB {0} [built]
    [3] ./ui/img/tip.png 297 bytes {0} [built]
    [4] ./ui/img/tip-locked.png 293 bytes {0} [built]
        + 2 hidden modules
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/treeherder-bugfiler.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/treeherder-bugfiler.css 303 bytes {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/treeherder-details-panel.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/treeherder-details-panel.css 15 KiB {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/treeherder-global.css:
    Entrypoint mini-css-extract-plugin = *
    [0] ../neutrino/node_modules/css-loader??ref--6-1!./ui/css/treeherder-global.css 8.5 KiB {0} [built]
        + 1 hidden module
Child mini-css-extract-plugin ../neutrino/node_modules/css-loader/index.js??ref--6-1!ui/css/treeherder-job-buttons.css:
    Entrypoint mini-css-extract-plugin = *
…